Completed Chapter 32
Completed Chapter 32: Multithreading and parallel programming. A very
hard topic to conceptualise and code. Debugging a parallel program was
quite difficult... note to self: test a recursive function outside of
the RecursiveAction/RecursiveTask to check if the function works before
implementing it (can save quite some time and prevent bugs)