Starting to use SQL
Had an issue downloading MySQL, but got it working so far. This Java
Book only contains command line for MySQL (and not using the workbench)
so will research more on MySQL. I will also research other  API's that
can be used instead of JDBC (which is used in the book).